pyproject.toml
[project]
name = "career-coach"
version = "0.1.0"
description = "Dockerized career coaching app with FastAPI and multi-agent Anthropic support"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
    "supabase>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
